backend/requirements.txt
# Dependências do Backend GuacPlayer
# Autor: GuacPlayer Team
# Data: 2025

# Framework Web
Flask==3.0.0
Flask-CORS==4.0.0

# Banco de Dados
psycopg2-binary==2.9.9

# Autenticação e Segurança
PyJWT==2.8.1

# Utilitários
python-dotenv==1.0.0
python-dateutil==2.8.2

# Logging e Monitoramento
python-json-logger==2.0.7

# Desenvolvimento e Testes
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
